% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphFunctions.R
\name{addEdge.adjListGraph}
\alias{addEdge.adjListGraph}
\title{A Graph Function}
\usage{
\method{addEdge}{adjListGraph}(graph, node1, node2, weight = 1)
}
\arguments{
\item{node1}{The first node}

\item{node2}{The second node}

\item{weight}{Defaults to 1 (used for non-weighted graph)}
}
\value{
the new graph with the added edge and whether the operation was successful
}
\description{
This function allows you to add a new edge to an existing graph. If weighted graph,
will override previous edge.
}
