% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphFunctions.R
\name{addEdge}
\alias{addEdge}
\title{A Graph Function}
\usage{
addEdge(node1, node2, weight = 1, bothDir = TRUE)
}
\arguments{
\item{node1}{the first node}

\item{node2}{the second node}

\item{weight}{Defaults to 1 (used for non-weighted graph)}

\item{bothDir}{Defaults to TRUE to add edge that goes  both ways (used for directionless graph)}
}
\value{
TRUE if the edge added or FALSE if edge already exists
addEdge()
}
\description{
This function allows you to add a new edge to an existing graph.
Throws error if 1 or both nodes do not exist.
}
