% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphFunctions.R
\name{addVertex.adjMatrixGraph}
\alias{addVertex.adjMatrixGraph}
\title{A Graph Function}
\usage{
\method{addVertex}{adjMatrixGraph}(graph, vertex)
}
\arguments{
\item{graph}{the graph obj}

\item{vertex}{the new vertex to add}
}
\value{
a list containing the new graph and whether an operation was performed
}
\description{
This function adds a vertex.
}
