% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphFunctions.R
\name{adjMatrixGraph}
\alias{adjMatrixGraph}
\title{A Graph Function}
\usage{
adjMatrixGraph(nodes, weighted = FALSE, directed = FALSE)
}
\arguments{
\item{nodes}{A vector of nodes}

\item{weighted}{TRUE if the graph  is weighted or FALSE. Defaults to FALSE.}

\item{directed}{TRUE if graph is directed. Defaults to FALSE.}
}
\value{
A graph object
}
\description{
This function returns a graph object stored as an adjacency matrix.
}
