% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphFunctions.R
\name{countEdges.adjListGraph}
\alias{countEdges.adjListGraph}
\title{A Graph Function}
\usage{
\method{countEdges}{adjListGraph}(graph)
}
\arguments{
\item{graph}{the graph obj}
}
\value{
a list containing the graph and the number of edges
}
\description{
This functions counts the edges.
}
