% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphFunctions.R
\name{countVertices.graph}
\alias{countVertices.graph}
\title{A Graph Function}
\usage{
\method{countVertices}{graph}(graph)
}
\arguments{
\item{graph}{the graph obj}
}
\value{
a list containing the graph and the number of vertices
}
\description{
This functions counts the vertices.
}
