\name{graph-defunct}
\alias{graph-defunct}
\title{Defunct Functions in Package \pkg{graph}}
%------ PLEASE: put \alias{.} here for  EACH !
\alias{pkgInstOrder}
\alias{buildRepDepGraph}
\alias{ugraphOld}
%
\description{
  The functions or variables listed here are no longer part of the graph package.
}
\usage{
buildRepDepGraph()
pkgInstOrder()
ugraphOld()
}
\seealso{
  \code{\link{Defunct}}
}


