% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphFunctions.R
\name{deleteEdge.adjMatrixGraph}
\alias{deleteEdge.adjMatrixGraph}
\title{A Graph Function}
\usage{
\method{deleteEdge}{adjMatrixGraph}(graph, node1, node2)
}
\arguments{
\item{graph}{the graph obj}

\item{node1}{The first node}

\item{node2}{The second node}
}
\value{
a list containing the new graph and whether an operation was performed
}
\description{
This function deletes an edge.
}
