% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphFunctions.R
\name{deleteVertex.adjListGraph}
\alias{deleteVertex.adjListGraph}
\title{A Graph Function}
\usage{
\method{deleteVertex}{adjListGraph}(graph, vertex)
}
\arguments{
\item{graph}{the graph obj}

\item{vertex}{the vertex to delete}
}
\value{
a list containing the new graph and whether an operation was performed
}
\description{
This function deletes a vertex.
}
