% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphFunctions.R
\name{graph}
\alias{graph}
\title{A Graph Function}
\usage{
graph(nodes, weighted = FALSE, directed = FALSE, useAdjMatrix = TRUE)
}
\arguments{
\item{nodes}{A vector of nodes}

\item{weighted}{TRUE if the graph  is weighted or FALSE. Defaults to FALSE.}

\item{directed}{TRUE if graph is directed. Defaults to FALSE.}

\item{useAdjMatrix}{TRUE if an adjacency matrix is used internally or FALSE to use a list}
}
\value{
A graph object
}
\description{
This function returns a graph object stored as an adjacency matrix.
}
