\name{graphExamples}
\alias{graphExamples}
\docType{data}
\title{A List Of Example Graphs}
\description{
This data set contains a list of example \code{graphNEL} objects, which can then
be used for plotting.
}
\usage{data(graphExamples)}
\source{
  Various sources, primarily from \code{\link{randomGraph}} and \code{link{randomEGraph}}
}
\examples{
data(graphExamples)
a <- graphExamples[[1]]
a
}
\keyword{datasets}
