% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphFunctions.R
\name{hasEdge.adjMatrixGraph}
\alias{hasEdge.adjMatrixGraph}
\title{A Graph Function}
\usage{
\method{hasEdge}{adjMatrixGraph}(graph, node1, node2)
}
\arguments{
\item{node1}{The first node}

\item{node2}{The second node}
}
\value{
a list containing the graph object and the a boolean indicating if the graph contained the edge
}
\description{
This function allows you to check for if an edge exists between 2 nodes in the graph.
}
