% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphFunctions.R
\name{hasEdge}
\alias{hasEdge}
\title{A Graph Function}
\usage{
hasEdge(node1, node2)
}
\arguments{
\item{node1}{the first node}

\item{node2}{the second node}
}
\value{
TRUE if the edge exists or FALSE otherwise
hasEdge()
}
\description{
This function allows you to check for if an edge exists between 2 nodes in the graph.
}
