% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram.R
\name{histogram}
\alias{histogram}
\title{Visualization of the distribution of a single variable using histograms}
\usage{
histogram(data, x, xlab = "x", bins = 30, na.rm = TRUE)
}
\arguments{
\item{data}{Default data set that supplies the data to generate the plot}

\item{x}{Column that contains a vector of numbers to plot}

\item{xlab}{Text that will be used for the title of the x axis}

\item{bins}{Specify the number of bins. Defaults to 30
could change this value, exploring multiple numbers of bins to find the best to
illustrate the stories in your data.}

\item{na.rm}{A logical vector indicating whether NA values
should be stripped before the computation proceeds. TRUE is the default.}
}
\value{
A histogram showing the distribution of x
}
\description{
This function displays the distribution of a single numeric variable
through a histogram
}
\examples{
library(gapminder)
histogram(gapminder, lifeExp)
histogram(gapminder, lifeExp, xlab = "Life Expectancy")
histogram(gapminder, gdpPercap, xlab = "GDP per capita", bins = 20)

}
