% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphFunctions.R
\name{isConnected.adjListGraph}
\alias{isConnected.adjListGraph}
\title{A Graph Function}
\usage{
\method{isConnected}{adjListGraph}(graph)
}
\arguments{
\item{graph}{the graph obj}
}
\value{
a list containing the graph and whether the graph is connected
}
\description{
This functions checks if the graph is connected.
}
