% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphFunctions.R
\name{isConnected.adjMatrixGraph}
\alias{isConnected.adjMatrixGraph}
\title{A Graph Function}
\usage{
\method{isConnected}{adjMatrixGraph}(graph)
}
\arguments{
\item{graph}{the graph obj}
}
\description{
This functions checks if the graph is connected.
}
