% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphFunctions.R
\name{isDense.adjMatrixGraph}
\alias{isDense.adjMatrixGraph}
\title{A Graph Function}
\usage{
\method{isDense}{adjMatrixGraph}(graph, level = 0.15)
}
\arguments{
\item{graph}{the graph obj}

\item{level}{the cutoff level for a dense graph, defaul is .85}
}
\value{
a list containing the graph and whether the graph is dense
}
\description{
This functions checks if the graph is dense.
}
