\name{isDirected-methods}
\docType{methods}
\alias{isDirected-methods}
\alias{isDirected}

\title{Determine if a graph has directed or undirected edges}
\usage{
isDirected(object)
}
\arguments{
  \item{object}{A \code{graph-class} instance}
}
\description{
  The edges of a \code{\link{graph-class}} object are either directed or
  undirected.  This function returns \code{TRUE} if the edges are
  directed and \code{FALSE} otherwise.
}
\keyword{methods}

