% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphFunctions.R
\name{isFullyConnected.adjListGraph}
\alias{isFullyConnected.adjListGraph}
\title{A Graph Function}
\usage{
\method{isFullyConnected}{adjListGraph}(graph)
}
\arguments{
\item{graph}{the graph obj}
}
\value{
a list containing the graph and whether the graph is fully connected
}
\description{
This functions checks if the graph is fully connected.
}
