% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphFunctions.R
\name{isFullyConnected.adjMatrixGraph}
\alias{isFullyConnected.adjMatrixGraph}
\title{A Graph Function}
\usage{
\method{isFullyConnected}{adjMatrixGraph}(graph)
}
\arguments{
\item{graph}{the graph obj}
}
\description{
This functions checks if the graph is fully connected.
}
