% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphFunctions.R
\name{isSparse.adjMatrixGraph}
\alias{isSparse.adjMatrixGraph}
\title{A Graph Function}
\usage{
\method{isSparse}{adjMatrixGraph}(graph, level = 0.15)
}
\arguments{
\item{graph}{the graph obj}

\item{level}{the cutoff level for a sparse graph, defaul is .15}
}
\value{
a list containing the graph and whether the graph is sparse
}
\description{
This functions checks if the graph is sparse.
}
