\name{MyGraph-package}
\alias{MyGraph-package}
\alias{MyGraph}
\docType{package}
\title{
  Package for work with Graph
}
\description{
  This package contains most famous graph algorithm 
}
\details{
  Allows you to search MST and lower path of weighted graph (direct or not direct) and create adjacency matrix of graph returned by traversal of entire graph or aql query.
}
\author{
Matteo, sabatinomatteo.98@gmail.com.

Maintainer: Matteo <sabatinomatteo.98@gmail.com>
}
\keyword{ package }

