\name{write.tlp}
\alias{write.tlp}
\title{ Write a graph object in a file in the Tulip format}
\description{
  Write a graph object in a file in the Tulip format.
}
\usage{
write.tlp(graph, filename)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{graph}{a \code{graph} object}
  \item{filename}{ Name of the output file}
}
\details{
  The Tulip format is used by the program Tulip.
}
\references{ About Tulip}
\author{ Laurent Gautier <laurent@cbs.dtu.dk>}
}
\keyword{ manip }

