\name{biocRepos}
\alias{biocRepos}
\docType{data}
\title{A graph representing the Bioconductor package repository}
\description{
  This graph is a rendition of the Bioconductor package repository and
  represents the dependency graph of that repository.  An edge between
  two package denotes a dependency on the 'to' package by the 'from'
  package.  
}
\usage{data(biocRepos)}
\examples{
   data(biocRepos)
   ## An example of usage will be here soon
}
\keyword{datasets}
